requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.22
Flask-Login==0.6.2
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.31.0
python-telegram-bot==20.4
Werkzeug==2.3.7
Flask_SQLAlchemy==3.0.2
